corrected setStop() TraCIAPI-method

+++ src/utils/traci/TraCIAPI.h
@@ -756,7 +756,7 @@ public:
         void setSpeed(const std::string& vehicleID, double speed) const;
         void setSpeedMode(const std::string& vehicleID, int mode) const;
         void setStop(const std::string vehicleID, const std::string edgeID, const double endPos=1.,
-                const int laneIndex=0, const int duration=std::numeric_limits<int>::max(),
+                const int laneIndex=0, const double duration=std::numeric_limits<double>::max(),
                 const int flags=0, const double startPos=std::numeric_limits<int>::min(),
                 const int until=-1) const;
         void setType(const std::string& vehicleID, const std::string& typeID) const;
